+++ src/api/caidapp/log_tools.py
@@ -14,3 +14,36 @@ class StatusCounts:
 
     def __str__(self):
         return str(self.counts)
+
+
+# filter which will count the number of same log messages
+class LogCounterFilter(logging.Filter):
+    def __init__(self):
+        super(LogCounterFilter, self).__init__()
+        self.status_counts = StatusCounts()
+
+    def filter(self, record):
+        self.status_counts.increment(record.getMessage())
+        return True
+
+    def __str__(self):
+        return str(self.status_counts)
+
+
+
+class DuplicateFilter(logging.Filter):
+    def __init__(self):
+        super().__init__()
+        self.last_log = None
+
+    def filter(self, record):
+        current_log = (record.module, record.levelno, record.msg)
+        if current_log != self.last_log:
+            self.last_log = current_log
+            return True
+        return False
+
+
+# Example
+# logger = logging.getLogger()
+# logger.addFilter(DuplicateFilter())